Quote argument with "[" to prevent shell globbing
Unix-style shells, including bash (which the ubuntu-latest CI
runner uses) and other POSIX-compatible shells, support a "["
notation similar to its meaning in regular expressions, as part of
their globbing (i.e., pattern matching) syntax.

Usually, "pip" commands that install a package with extras without
suppressing the special meaning of "[" by quoting it nonetheless
succeed, including this case (so far). This is because, when no
existing entry matches the glob pattern, the default behavior (in
bash, and most other shells) is to fall back to treating the glob
pattern literally. However, new files/directories could break this,
which would not be intended. (Quoting the argument in the shell
command also has the benefit of unambiguously expressing intent.)

https://www.gnu.org/software/bash/manual/bash.html#Pattern-Matching-1
https://pubs.opengroup.org/onlinepubs/9699919799.2018edition/utilities/V3_chap02.html#tag_18_13