Revert "Fix unit test" to observe failure
The bug, whose fix this reverts, is in the code under test. I am
reverting the fix on this branch to verify that the tests do find
the bug, on all Python versions and platforms tested. (So I may
merge this into a child branch that that has multiple operating
systems in the CI test matrix.)

Note that the bug is only observed when pandas is installed. That
happens on CI due to the extras that are included. (Conceptually,
datalib is the extra that ensures pandas is installed for the
purpose of the code under test that has the bug. However, some of
the other extras also specify it.) This may be why the bug had not
been detected initially, but I want to verify that it causes a test
failure on many Python versions and platforms, in case there are
any other subtleties that might merit attention.

This reverts commit c46adb5ab4814330e9db5387fc70fc20dff99b48.